eval on find refs to that arrows resolve
