+++ server/src/symbols/methodSymbol.ts
@@ -26,10 +26,8 @@ import {
 import { DiagnosticSeverity, FoldingRange } from "vscode-languageserver";
 import { ExpressionSymbol } from "./expressionSymbol";
 import { lexRangeFromContext, rangeFromTokens } from "../utils";
-import { ObjectReferenceInfo } from "./objectSymbol";
 import { ContextSymbolTable } from "../backend/ContextSymbolTable";
 import {
-    ArrayStackValue,
     CallStack,
     RootFrame,
     StackFrame,
@@ -42,6 +40,7 @@ import { InheritSuperAccessorSymbol } from "./inheritSymbol";
 import { resolveOfTypeSync } from "../backend/symbol-utils";
 import { asStackValue } from "../backend/CallStackUtils";
 import { MethodParameterSymbol } from "./variableSymbol";
+import { ArrowSymbol } from "./arrowSymbol";
 
 export const MAX_CALLDEPTH_SIZE = 25;
 
@@ -172,110 +171,6 @@ export class MethodDeclarationSymbol
     }
 }
 
-export class MethodInvocationSymbol
-    extends ScopedSymbol
-    implements IEvaluatableSymbol, IKindSymbol
-{
-    public methodName: string;
-    public methodSymbol: MethodSymbol;
-
-    public getArguments() {
-        return getSymbolsOfTypeSync(this, ExpressionSymbol);
-    }
-
-    public get kind() {
-        return SymbolKind.MethodInvocation;
-    }
-
-    constructor(name: string) {
-        super(name);
-    }
-
-    eval(stack: CallStack, callScope?: RootFrame) {
-        // if a callScope was passed, that stack frame is used for fn lookup
-        const fnLookupFrame = callScope ?? stack.getCurrentRoot();
-
-        // find the function that this invocation points to
-        const funcIdValue = stack.getValue<string>(FUNCTION_NAME_KEY, this);
-        stack.clearValue(FUNCTION_NAME_KEY);
-        const methodName = (this.methodName = funcIdValue?.value);
-
-        let methodSymbol: MethodSymbol;
-
-        const superAcc = this.getParentOfType(InheritSuperAccessorSymbol);
-        if (!!superAcc) {
-            // this is a special case where we need to get the method from the super accessor's symbol table
-            // try actual method first, then definitions & efuns
-
-            // NTBLA: This is actualy incorrectly - each inherit should add a frame to the stack so that we
-            // can just walk up the stack and look for the inherited method.
-
-            methodSymbol = resolveOfTypeSync(
-                superAcc.objSymbolTable,
-                methodName,
-                MethodSymbol,
-                false
-            );
-            methodSymbol ??= resolveOfTypeSync(
-                superAcc.objSymbolTable,
-                methodName,
-                LpcBaseMethodSymbol,
-                false
-            );
-            this.methodSymbol = methodSymbol;
-        } else {
-            // just get method out of stack
-            methodSymbol = this.methodSymbol = getFunctionFromFrame(
-                fnLookupFrame,
-                methodName
-            );
-        }
-
-        if (!funcIdValue) {
-            const ctx = this.context as ParserRuleContext;
-            addDiagnostic(this, {
-                message: `Function could not be validated`,
-                range: rangeFromTokens(ctx.start, ctx.stop),
-                type: DiagnosticSeverity.Hint,
-                code: DiagnosticCodes.FunctionUnknown,
-            });
-        } else if (!methodSymbol) {
-            const ctx = (funcIdValue?.symbol ?? this)
-                .context as ParserRuleContext;
-            // if (methodName == "call_other") {
-            //     const ii = 0;
-            // }
-            addDiagnostic(this, {
-                message: `Cannot find function named '${methodName ?? ""}'.`,
-                range: rangeFromTokens(ctx.start, ctx.stop),
-                type: DiagnosticSeverity.Warning,
-                code: DiagnosticCodes.FunctionNotFound,
-            });
-        }
-
-        const prms = this.getArguments().filter((c) =>
-            isInstanceOfIEvaluatableSymbol(c)
-        ) as IEvaluatableSymbol[];
-        // for (const child of this.children) {
-        //     if (isInstanceOfIEvaluatableSymbol(child)) {
-        //         scope = child.eval(stack, scope);
-        //     } else {
-        //         console.warn("not evaluable: " + child.name);
-        //     }
-        // }
-
-        return methodSymbol?.eval(stack, prms, callScope);
-    }
-
-    /**
-     * get the method symbol that this invocation points to.
-     */
-    public getMethodSymbol() {
-        // find the context that this method is being invoked on
-        return this.methodSymbol;
-    }
-}
-
 /**
  * Represents a code identifier that points to a function somehwere else, such as `foo()` or `o->bar()`.
  * On evaluation, it will attempt to look up the actual function definition and evaluate it.
@@ -295,8 +190,14 @@ export class FunctionIdentifierSymbol
     getReference(): BaseSymbol {
         if (!this.reference) {
             // try to fill in the ref if it wasn't already identified
+            let table: ContextSymbolTable;
+            if (this.parent instanceof ArrowSymbol) {
+                table = (this.parent as ArrowSymbol).objContext?.symbolTable;
+            }
+            table &&= this.symbolTable as ContextSymbolTable;
+
             this.reference = resolveOfTypeSync(
-                this.symbolTable,
+                table,
                 this.name,
                 LpcBaseMethodSymbol,
                 false
